vmm: defer MMIO instruction fetch
Fetch VMX instruction bytes only after vm_fault rejects the GPA. Return an executable prefix so a valid instruction at a page boundary does not depend on the following page.

Keep SVM decode-assist data intact and move KVM tests to the installed sys/kvm.h ABI header.

+++ sys/dev/virtual/vmm/x64/vmm_x64_emul.h
@@ -18,6 +18,7 @@ int vmm_x64_emul_resume(struct vmm_vcpu *);
 int vmm_x64_emul_complete_write(struct vmm_vcpu *);
 int vmm_x64_emul_complete_read(struct vmm_vcpu *, const void *, size_t);
 int vmm_x64_translate(struct vmm_vcpu *, uint64_t, uint64_t *);
-int vmm_x64_fetch_instruction(struct vmm_vcpu *, uint8_t *, size_t);
+/* Returns the contiguous executable prefix fetched from guest RIP. */
+size_t vmm_x64_fetch_instruction(struct vmm_vcpu *, uint8_t *, size_t);
 
 #endif /* VMM_X64_EMUL_H */
